% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformar.R
\name{a_caracteres}
\alias{a_caracteres}
\title{Convertir columnas de un data frame a caracteres}
\usage{
a_caracteres(df, columnas, ...)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
Convertir columnas de un data frame a caracteres
}
\examples{
# Las columnas se pueden ingresar por su índice...
mtcars <- a_caracteres(df = mtcars, columnas = 3:7)

# O también por el nombre:
mtcars <- a_caracteres(df = mtcars, columnas = c("hp", "wt", "am"))

}
