% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformar.R
\name{a_cero}
\alias{a_cero}
\title{Convertir los NA a ceros en columnas de data frame}
\usage{
a_cero(df, columnas)
}
\arguments{
\item{columnas}{}
}
\value{

}
\description{
Convertir los NA a ceros en columnas de data frame
}
