% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformar.R
\name{a_nas}
\alias{a_nas}
\title{Convertir los valores de las columnas seleccionadas a NAs}
\usage{
a_nas(df, columnas)
}
\arguments{
\item{df}{\strong{Un data frame}. La función admite sólo data frames, no objetos
derivados, como tibles.}

\item{columnas}{\strong{Un vector con las columnas a transformar}. Puede ser
vector numérico con el índice de las columnas (\code{columnas = 2:4}) o
vector de caracteres con el nombre de las columnas
(\verb{columnas = c("a", "b"})).}
}
\value{
Un dataframe con los valores de las columnas del argumento
convertidas a NAs.
}
\description{
Convierte los valores de las columnas de un data frame a
NAs.
}
\details{
En principio, no es relevante para la función si las columnas
seleccionadas son numéricas, de caracteres o lógicas. El resultado será
siempre todos sus valores convertidos a NAs.
}
\examples{
datos <- data.frame(a = c(10,15,20, 25), b = c(10, 20, 30, 40))
(datos <- a_nas(datos, 1))
}
\seealso{
Other funciones para transformar columnas de data frames: 
\code{\link{a_caracteres}()},
\code{\link{a_cero}()},
\code{\link{a_numeros}()},
\code{\link{formatear_num}()}
}
\concept{funciones para transformar columnas de data frames}
