% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basicas.R
\name{\%notin\%}
\alias{\%notin\%}
\title{¿No está en un vector?}
\usage{
x \%notin\% y
}
\arguments{
\item{x}{\strong{Valor o elemento numérico o de caracter}. Es el que se requiere
determinar si se encuentra en el vector \code{y}.}

\item{y}{\strong{Vector}. Es al que se pregunta si no contiene el valor o elemento
\code{x}.}
}
\value{
Valor lógico \strong{\code{TRUE}} o \strong{\code{FALSE}}.
}
\description{
Determina si un valor o elemento dado \strong{no se encuentra} en un
vector. Es la negación de la función \code{\link{\%in\%}}. La
función \code{\%in\%} arroja \code{TRUE} si el elemento está en el vector. En cambio,
esta función \verb{\%notin\%} arroja \code{TRUE} si el elemento \strong{no se encuentra} en
el vector.
}
\details{
El paquete \code{Hmisc} contiene la función \code{\link[Hmisc]{\%nin\%}}
que hace la misma operación: determinar si un valor o elemento no se
encuentra en un vector. Su nombre, sin embargo, es poco intuitivo y se
demora el doble de tiempo en ejecutar la misma operación.
}
\examples{
# Para saber si un valor o elemento de una función **si está** en un vector:
  "a" \%in\% letters # Resultado: TRUE, dado que "a" si está en el vector.

# La función `\%notin\%` arroja el resultado contrario:
  "a" \%notin\% letters # Resultado: FALSE

# Al igual que `\%in\%`, `\%notin\%` tiene gran utilidad en el control de flujo:
  dias <- weekdays(x=as.Date(seq(6), origin="1950-01-01"))
  if ("domingo" \%notin\% dias) dias <- c(dias, "domingo")
}
