% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basicas.R
\name{que.es}
\alias{que.es}
\title{Determinar qué tipo de objeto con precisión}
\usage{
que.es(x, ...)
}
\arguments{
\item{...}{}
}
\value{
vector
}
\description{
Determinar qué tipo de objeto con precisión
}
\examples{
bbdd <- data(mtcars)
que.es(bbdd)
}
