% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformar.R
\name{a_numeros}
\alias{a_numeros}
\title{Convertir columnas de un data frame a números}
\usage{
a_numeros(df, columnas, ...)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
Convertir columnas de un data frame a números
}
\examples{
# Teniendo un data frame con columnas que tienen números pero, por diversas
 circunstancias, como vectores de caracteres....
cuadro = data.frame (a = as.character(c(1:5)), b = as.character(c(6:10)))
# a_numeros generará
# cuadro <- a_numeros(cuadro, 1:2)
}
