% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basicas.R
\name{listar}
\alias{listar}
\title{Incorporar objetos a una lista}
\usage{
listar(lista, obj, pos = 0, rm = T)
}
\arguments{
\item{lista}{\strong{Una lista}. Si la lista no existe en el entorno global
(GlobalEnv), la función creará una y le asignará el obejto.}

\item{obj}{\strong{Objeto de cualquier tipo}. Es necesario que exista previamente en
el entorno global (GlobalEnv).}

\item{pos}{\strong{Un número}. Puede asumir dos valores:
\itemize{
\item 0: la fila se añade al final la lista (por defecto);
\item 1: la fila se añade al principio de la lista;
}}

\item{rm}{\strong{Valor lógico}. Si es \code{TRUE} (\emph{predeterminado}), el objeto será eliminado del entorno
global (GlobalEnvir) después de ser agregado a la lista. Si es \code{FALSE},
el objeto será añadido a la lista sin ser eliminado del entorno global.}
}
\value{
Una \emph{\code{lista}} con el \emph{\code{obj}} incorporado en la posición indicada con el
argumento \emph{\code{pos}}.
}
\description{
Agregar objetos a una lista de forma rápida controlando la
posición y qué pasa con el objeto en el entorno global (GlobalEnv).
}
\details{
Si \code{obj} no existe antes de aplicar la función, arrojará error. Debe
existir antes en el entorno global.

Si \code{listar()} se usa dentro de una función personalizada, debe definirse
\code{rm = FALSE}. De lo contrario, arrojará error.

Si \code{listar()} se usa múltiples veces en un pipe de \code{magrittr} (el \verb{\%>\%}), sólo
añadirá el último objeto de la secuencia.
}
\examples{

# Creamos primero una lista a la después se le agregarán elementos.

mis_cosas <- list(
   libros = c("El péndulo de Foucault", "Germinal", "Artificios", "Angela's Ashes"),
   vinilos = c("Abbey Road", "Kind of blue", "Nevermind",
    "Bigger, better, faster, more", "Parte de la religión")
  )

 # Creamos después elementos para agregar a la lista

computadoras <- c("Notebook", "Workstation", "Raspberry Pi")

# Agregando el objeto al final de la lista y conservándolo en el entorno global:
  mis_cosas <- listar(mis_cosas, computadoras, pos = 0, rm = FALSE)

# Agregando el objeto al primer lugar de la lista y eliminándolo del entorno global:
  mis_cosas <- listar(mis_cosas, computadoras, pos = 1, rm = TRUE)
}
